+++ src/pages/admin/evaluation-administrations/[id]/evaluees/evaluees.tsx
@@ -1,7 +1,7 @@
 import { EvalueesList } from "../../../../../features/admin/evaluation-administrations/[id]/evaluees/evaluees-list"
-import { EvalueesFilter } from "../../../../../features/admin/evaluation-administrations/[id]/evaluees/EvalueesFilter"
-import { EvalueesFooter } from "../../../../../features/admin/evaluation-administrations/[id]/evaluees/EvalueesFooter"
-import { EvalueesHeader } from "../../../../../features/admin/evaluation-administrations/[id]/evaluees/EvalueesHeader"
+import { EvalueesFilter } from "../../../../../features/admin/evaluation-administrations/[id]/evaluees/evaluees-filter"
+import { EvalueesFooter } from "../../../../../features/admin/evaluation-administrations/[id]/evaluees/evaluees-footer"
+import { EvalueesHeader } from "../../../../../features/admin/evaluation-administrations/[id]/evaluees/evaluees-header"
 
 export default function Evaluees() {
   return (
